requirements.txt
Django==1.9.7
Markdown==2.6.6
django-reversion==2.0.5
djangorestframework==3.3.3
wsgiref==0.1.2
